# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/blob/{key}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Get blob data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}